+++ packages/mixins/crud.js
@@ -73,32 +73,10 @@ export default function() {
             this.init();
         },
         methods: {
-            columnSort() {
-                var compare = function(prop) {
-                    return function(obj1, obj2) {
-                        var val1 = obj1[prop];
-                        var val2 = obj2[prop];
-                        if (!isNaN(Number(val1)) && !isNaN(Number(val2))) {
-                            val1 = Number(val1);
-                            val2 = Number(val2);
-                        }
-                        if (val1 < val2) {
-                            return -1;
-                        } else if (val1 > val2) {
-                            return 1;
-                        } else {
-                            return 0;
-                        }
-                    }
-                }
-                this.tableOption.column.sort(compare('sort'));
-            },
             init() {
                 // 初始化工具
                 this.initFun();
                 this.tableOption = this.deepClone(this.option);
-                //字段排序
-                this.columnSort();
                 const dicFlag = this.vaildData(this.tableOption.dicFlag, true);
                 // 规则初始化
                 this.rulesInit();